requirements.txt
glfw
pyopengl
pyrr
pillow
numpy
pywavefront